% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simu.r
\docType{data}
\name{simu}
\alias{simu}
\title{Simulated latent class data with competing risks}
\format{
a data frame of 3 latent classes (with unknown label), with 500 subjects, 2 baseline covariates (Y.1 and Y.2), time of competing risks (ftime), and failure types (fstatus). Failure types include type 1, 2, censored 0, or missing NA.
}
\usage{
data(simu)
}
\description{
Simulated latent class data with competing risks
}
\keyword{datasets}
