% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.r
\name{simulation}
\alias{simulation}
\title{Simulate an example dataset}
\usage{
simulation(n, p.class, mu, sigma, lambda, p.cif, cl, cu)
}
\arguments{
\item{n}{Sample size}

\item{p.class}{L-dimensional marginal probability vector of latent classes. L is the number of latent classes.}

\item{mu}{Latent class-specific mean vectors for baseline covariates, stored in a matrix of dimension L by n. The ith column of the matrix is the mean for the ith latent class.}

\item{sigma}{Latent class-specific covariance matrices for baseline covariates, stored in a list. See example for detailed specifications.}

\item{lambda}{Underlying true latent class effect coefficients, vector of L}

\item{p.cif}{A scalar parameter to adjust the subdistribution survival function under Fine and Gray (1999)'s model.}

\item{cl}{Lower bound of censoring time.}

\item{cu}{Upper bound of censoring time.}
}
\value{
Returns a data frame with event time, failure type and baseline covariates. Failure type 0 indicates censoring. Missing failure types for uncensored subjects are denoted by NAs.
}
\description{
Simulate an example dataset.
}
\details{
Simulate latent classes from multinomial distribution, then generate baseline covariates based on finite mixture model (McLachlan and Peel, 2000) and competing risks data with 2 failure types based on Fine and Gray (1999)'s proportional subdistribution hazards model. Censoring is independently generated by uniform distribution (cl, cu). For uncensored subjects, missingness of failure type is independently generated from a logit function depending on baseline covariates and failure time.
}
\examples{

#The following example specifies all required parameters and generate a dataset.
gamma <- c(0.5,0.5,-1)
pi = c(0.3,0.35,0.35)
mu = matrix(c(1,1,2.5,2.5,4,4),nrow=2,ncol=3)
sigma1 = matrix(c(0.36,0.27,0.27,0.81),2,2)
sigma2 = matrix(c(0.49,0.504,0.504,0.64),2,2)
sigma3 = matrix(c(0.25,0.225,0.225,0.25),2,2)
sigma = list(sigma1,sigma2,sigma3)
p.cif = 0.66
cl=0.19
cu=1.09
dat=raw(500,pi,mu,sigma,gamma,p.cif,cl,cu)
}
\references{
Fine and Gray (1999). A proportional hazards model for the subdistribution of a competing risk. Journal of the American statistical association 94, 496-509.\cr
McLachlan and Peel (2000). Finite Mixture Models. John Wiley & Sons.
}
\author{
Teng Fei. Email: tfei@emory.edu
}
